[Modal] Prevent backdrop to stay open (#16694)
* Modal: prevent backdrop to stay open when 'open' is flickering

This should address #12831.

After following the comments in the above issue, I indeed noticed that `onExited` was not called. However, I've also noticed that `onEntered` was not called either. This is because the transition is not even rendered.

To solve this, I have added a flag that tells if the transition was entered. If this is not the case, we can remove the modal from the DOM when open is false.

* [Modal] Fix exited state